Refactor: isolated class ProductionPlanCalculator
